feat: add day 25 interactive
Doesn't seem worth the effort to write code that navigates the ship,
picks up only the safe items (which were determined by trial and error),
then selects the combo required to open the door. The last part of that
(selecting items to open the door) is certainly algorithmic, and I
might've implemented it if I didn't have to wade through the rest of the
pile to do so. Just not worthwhile.